Add cache package and implement cache functionality.
This commit adds a new package called "cache" and implements various cache functionality. The cache package is used to store and retrieve data in memory with expiration times.

The cache package includes the following features:
- The Item struct defines an object to be stored in the cache, along with its expiration time.
- The Cache struct provides a high-level interface to interact with the cache.
- The cache struct contains the actual cache data structure, as well as methods to manipulate it.
- The methods in the cache struct allow setting, getting, and deleting items from the cache, as well as setting default expiration times.

The cache package can be used to store any type of object in memory and retrieve it later. Each item in the cache has an associated expiration time, which can be used to automatically remove stale data from the cache.

This commit introduces the cache package and provides the basic functionality needed to store and retrieve data in the cache.